Handling new docker compose command & commented out TG since licensing is an issue now
