+++ drivers/neo4j/src/test/resources/docker/Neo4j.yml
@@ -1,4 +1,4 @@
-version: '3'
+
 services:
   neo4j:
     image: bitnami/neo4j:4.3.6
