+++ drivers/neo4j/src/test/scala/com/github/plume/oss/drivers/Neo4jIntTests.scala
@@ -5,6 +5,8 @@ import com.github.plume.oss.testfixtures.PlumeDriverFixture
 
 class Neo4jIntTests extends PlumeDriverFixture(new Neo4jDriver()) {
 
+  private implicit val loader: ClassLoader = getClass.getClassLoader
+
   override def beforeAll(): Unit = {
     DockerManager.startDockerFile("Neo4j", List("plume-neo4j"))
     driver.asInstanceOf[IDriver & ISchemaSafeDriver].buildSchema()
