--- drivers/tigergraph/src/test/resources/docker/Dockerfile
@@ -1,3 +1,0 @@
-FROM tigergraph/tigergraph:3.9.3-1
-ENTRYPOINT /usr/sbin/sshd && \
-    su - tigergraph bash -c "/home/tigergraph/tigergraph/app/cmd/gadmin start all && tail -f /dev/null"
