+++ project/Versions.scala
@@ -2,24 +2,26 @@ object Versions {
 
   // SAST
   val codePropertyGraph = "1.6.11"
-  val joern             = "2.0.379"
-  val overflowDb        = "1.192"
+  val joern = "2.0.379"
 
   // Drivers
   val tinkerGraph = "3.4.11"
-  val neo4j       = "4.4.5"
+  val neo4j = "4.4.5"
+  val overflowDb = "1.192"
+  val flatGraph = "0.0.59"
 
   // Utilities
   val apacheCodec = "1.15"
-  val apacheIo    = "2.11.0"
-  val apacheLang  = "3.12.0"
-  val sttp        = "3.9.0"
-  val jackson     = "2.13.2"
-  val lz4         = "1.8.0"
-  val slf4j       = "2.0.5"
-  val log4j       = "2.20.0"
-  val logback     = "1.2.11"
-  val scalatest   = "3.2.15"
-  val circe       = "0.14.1"
+  val apacheIo = "2.11.0"
+  val apacheLang = "3.12.0"
+  val apacheCommonsText = "1.12.0"
+  val sttp = "3.9.0"
+  val jackson = "2.13.2"
+  val lz4 = "1.8.0"
+  val slf4j = "2.0.5"
+  val log4j = "2.20.0"
+  val logback = "1.2.11"
+  val scalatest = "3.2.15"
+  val circe = "0.14.1"
 
 }
